# src/main/resources/doc_swagger.yaml
swagger: '2.0'
info:
  description: >-
    Sugestão de documentação para ser utilizada como base para o desenvolvimento
    da API do projeto OPEN PMO da SEP-ES. 
  version: 1.0.0
  title: Open PMO Api
  termsOfService: 'http://www.totemti.com.br/'
  contact:
    email: atendimento@totemti.com.br
host: localhost
basePath: /openPmo
...
schemes:
  - https
  - http
paths:
...
  /offices:
    get:
      tags:
        - officeController
      summary: Get all office
      description: ''
      operationId: getAllOffice
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/OfficeDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - officeController
      summary: Add a new office
      description: ''
      operationId: addOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OfficeParamDto'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    put:
      tags:
        - officeController
      summary: Update an office
      description: ''
      operationId: updateOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OfficeParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error          
  '/offices/{id}':
    get:
      tags:
        - officeController
      summary: Get office
      description: ''
      operationId: getOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OfficeDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    
    delete:
      tags:
        - officeController
      summary: Delete an office
      description: ''
      operationId: deleteOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
  /office-permissions:
    get:
      tags:
        - officePermissionController
      summary: Get all office permissions
      description: ''
      operationId: getAllPermissionOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: id
          description: id office
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OfficePermissionDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - officePermissionController
      summary: Add a new office permissions
      description: ''
      operationId: postPermissionOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OfficePermissionParamDto'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    put:
      tags:
        - officePermissionController
      summary: Update an office
      description: ''
      operationId: officePermissionUpdate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/OfficePermissionParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error          
...
  /localities:
    get:
      tags:
        - localityController
      summary: Get all localities from a domain
      description: ''
      operationId: getAllLocality
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: id-domain
          description: id from domain
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/LocalityDto'
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    post:
      tags:
        - localityController
      summary: Add a new Locality to a domain
      description: ''
      operationId: addLocalityDomain
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/LocalityParamDto'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/EntityDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    put:
      tags:
        - localityController
      summary: Update a locality from a domain
      description: ''
      operationId: updateLocality
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/LocalityParamDto'
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
  '/localities/{id}':
    get:
      tags:
        - localityController
      summary: Get locality from a domain
      description: ''
      operationId: getLocality
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from locality
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/LocalityDetailDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - localityController
      summary: Delete a locality from a domain
      description: ''
      operationId: deleteLocality
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from locality
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  EntityDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
...
  OfficeDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
  OfficeParamDto:
    type: object
    required:
      - name
      - fullName
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
  OfficePermissionDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idPerson:
        type: integer
        format: int64
      permissionLevel:
        type: string
      permittedRole:
        type: string
      organization:
        type: string
  OfficePermissionParamDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idOffice:
        type: integer
        format: int64
      idPerson:
        type: integer
        format: int64
      permissionLevel:
        type: string
      permittedRole:
        type: string
      organization:
        type: string
...
  DomainDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
...
  LocalityDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      domain:
        $ref: '#/definitions/DomainDto'
      children:
        $ref: '#/definitions/LocalityChildrenDto'  
      type:
        type: string
  LocalityChildrenDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      type:
        type: string
  LocalityDetailDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idDomain:
        type: integer
        format: int64
      idParent:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
      latitude:
        type: string
      longitude:
        type: string
  LocalityParamDto:
    type: object
    required:
      - name
      - fullName
    properties:
      id:
        type: integer
        format: int64
      idDomain:
        type: integer
        format: int64
      idParent:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
      latitude:
        type: string
      longitude:
        type: string
...